feat: json everywhere
BREAKING CHANGE: The POST and PUT methods now return JSON by default, with the JWT located in the `token` field. The response can return just the token if `application/jwt` is set in the `Accept` header. All HTTP errors are now of JSON type, with `error: true` (as opposed to valid requests having `ok: true`)